requirements.txt
appworld
ollama
requests

numpy
sentence-transformers
python-dotenv